3B2 cppcheck fixes
- Variable scope reduction
- Use typedefs for rapp_data and capp_data to facilitate better bounds
  checks.
- Actual bound check on cio in cio_requeue_avail: cppcheck seems to
  think that the cio array should be 14 elements in size. Add the bound
  check just in case cppcheck is correct.

Non-cppcheck fixes:
- Use size_t for loop induction variables when iterating through arrays
  when it makes sense.

+++ 3B2/3b2_ni.h
@@ -184,7 +184,7 @@ typedef struct {
     ETH_PACK        rd_buf;
     ETH_PACK        wr_buf;
     ETH_MAC         macs[NI_FILTER_MAX];    /* List of all filter addresses */
-    int             filter_count;           /* Number of filters available */
+    size_t          filter_count;           /* Number of filters available */
     ETH_PCALLBACK   callback;
 } NI_STATE;
